Refactor, split dock from model
Don't show dock on statup, but start logging immediately. Lazily
create the dock widget only on first F12 press.

TODO - add toolbar/menu entry for more obvious dock display,
for people who aren't Richard or me?
